backend/python/requirements.txt
paho-mqtt
xlrd==1.2.0